Promote cpp_eve_item_id to eve_items primary key, drop auto-increment id, and drop dependent views before column widening to unblock the PostgreSQL migration.

+++ evebs/models/tables/associations.py
@@ -3,7 +3,7 @@ from evebs.extensions import db
 eve_items_users = db.Table('eve_items_users',
     db.Column('id', db.Integer, primary_key=True),
     db.Column('user_id', db.Integer, db.ForeignKey('users.id')),
-    db.Column('eve_item_id', db.Integer, db.ForeignKey('eve_items.id')),
+    db.Column('eve_item_id', db.BigInteger, db.ForeignKey('eve_items.id')),
 )
 
 trade_hubs_users = db.Table('trade_hubs_users',
